+++ KokeiroLifeLogger/Repositories/ThermohygrometerRepository.cs
@@ -10,9 +10,9 @@ namespace KokeiroLifeLogger.Repositories
     {
         public string Location { get; set; }
 
-        public int Temperature { get; set; }
+        public double Temperature { get; set; }
 
-        public int Humidity { get; set; }
+        public double Humidity { get; set; }
 
         public DateTime MesuredAt { get; internal set; }
